change guarantees section layout and split TransactionAccountCard into smaller composables
